requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
loguru==0.7.0
Pillow==9.5.0
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.6.3
uvicorn==0.22.0
